Update URLs to HTTPS and Clean up Unused Code (#28)
* Remove unused CLIENT_SUBDOMAIN_HEADER constant

The CLIENT_SUBDOMAIN_HEADER constant was no longer being used and has been removed to clean up unused code. This helps improve readability and maintainability of the class.

* Remove unused request header validation logic.

Deleted the `RequestHeaders` decorator, `RequestingUserEmailHeaderDto` class, and `supabaseUserIdKey` constant as they are no longer in use. This cleanup reduces unused code and improves maintainability of the codebase.

* Refactor Match validator to remove unused parameter.

Renamed the 'type' parameter to '_type' to indicate it is unused. This improves code clarity and aligns with convention for unused variables. No functional changes were made.

* Use `getOrThrow` for mandatory DATABASE_URL retrieval

Replaced `get` with `getOrThrow` to ensure the `DATABASE_URL` environment variable is mandatory and throws an error if not provided. This change improves reliability and safeguards against missing configurations during Prisma migrations.

* Update widget script URL to use HTTPS

Switched the widget script URL from HTTP to HTTPS to ensure secure loading. This change improves security and avoids potential issues with mixed content on secure websites.